Finished converting to a child_process.spawn() implementation with promisified stdout callback functions.
Known problem: The buffer can end at any time so in onP4Changes you might get a partial line at the start of a new chunk of stdout. We need a way to save the last line of a chunk for later re-stitching with the first line of a new chunk.