WIP - Still trying to debug the promise chain.
Some of the promises resolve before logging but we don't seem to be awaiting the resolution of ALL of them, and I'm not sure why yet.

I was using Promise.all() right before the return of each function and then Promise.all() again on a series of Promises produced by Promise.all(), which I would expect should result in a single promise that resolves when all of the subpromises resolve but it doesn't seem to be working that way.